pyproject.toml
[project]
name = "crm-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "gunicorn>=23.0.0",
    "logger>=1.4",
    "neo4j>=6.0.2",
    "pydantic[email]>=2.12.3",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.38.0",
]
